refactor: reorganization and docstrings
