--- exam_project/customer_classification_model/utils.py
@@ -1,76 +1,0 @@
-import os
-import time
-
-from mlflow.entities.model_registry.model_version_status import ModelVersionStatus
-from mlflow.tracking.client import MlflowClient
-import pandas as pd
-
-
-def describe_numeric_col(x):
-    """
-    Parameters:
-        x (pd.Series): Pandas col to describe.
-    Output:
-        y (pd.Series): Pandas series with descriptive stats.
-    """
-    return pd.Series(
-        [x.count(), x.isnull().count(), x.mean(), x.min(), x.max()],
-        index=["Count", "Missing", "Mean", "Min", "Max"],
-    )
-
-
-def impute_missing_values(x, method="mean"):
-    """
-    Parameters:
-        x (pd.Series): Pandas col to describe.
-        method (str): Values: "mean", "median"
-    """
-    if (x.dtype == "float64") | (x.dtype == "int64"):
-        x = x.fillna(x.mean()) if method == "mean" else x.fillna(x.median())
-    else:
-        x = x.fillna(x.mode()[0])
-    return x
-
-
-def create_dummy_cols(df, col):
-    df_dummies = pd.get_dummies(df[col], prefix=col, drop_first=True)
-    new_df = pd.concat([df, df_dummies], axis=1)
-    new_df = new_df.drop(col, axis=1)
-    return new_df
-
-
-def wait_until_ready(model_name, model_version):
-    client = MlflowClient()
-    for _ in range(10):
-        model_version_details = client.get_model_version(
-            name=model_name,
-            version=model_version,
-        )
-        status = ModelVersionStatus.from_string(model_version_details.status)
-        print(f"Model status: {ModelVersionStatus.to_string(status)}")
-        if status == ModelVersionStatus.READY:
-            break
-        time.sleep(1)
-
-
-def wait_for_deployment(model_name, model_version, stage="Staging"):
-    client = MlflowClient()
-    status = False
-    while not status:
-        model_version_details = dict(
-            client.get_model_version(name=model_name, version=model_version)
-        )
-        if model_version_details["current_stage"] == stage:
-            print(f"Transition completed to {stage}")
-            status = True
-            break
-        else:
-            time.sleep(2)
-    return status
-
-
-def create_directories():
-    os.makedirs("artifacts", exist_ok=True)
-    os.makedirs("mlruns", exist_ok=True)
-    os.makedirs("mlruns/.trash", exist_ok=True)
-    print("Created artifact and mlruns directories")
